Made cool auto scrolling!!
Added auto scrolling to the page so that when you expand the accordion
it scrolls to the bottom. When you un-expand the accordion it scrolls
back to the top. Might be a problem if the accordion text gets too
long. Also added another image to the home screen. Will have to change
it probably. The text in the captions isn’t very easy to read.